third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/registration-updateviacache.https-expected.txt

This is a testharness.js-based test.
FAIL register-with-updateViaCache-undefined assert_true: New worker installing expected true got false
FAIL register-with-updateViaCache-imports assert_true: New worker installing expected true got false
FAIL register-with-updateViaCache-all assert_equals: reg.updateViaCache expected "all" but got "imports"
FAIL register-with-updateViaCache-none assert_equals: reg.updateViaCache expected "none" but got "imports"
PASS register-with-updateViaCache-undefined-then-undefined
PASS register-with-updateViaCache-undefined-then-imports
FAIL register-with-updateViaCache-undefined-then-all assert_equals: reg.updateViaCache updated expected "all" but got "imports"
FAIL register-with-updateViaCache-undefined-then-none assert_equals: reg.updateViaCache updated expected "none" but got "imports"
PASS register-with-updateViaCache-imports-then-undefined
PASS register-with-updateViaCache-imports-then-imports
FAIL register-with-updateViaCache-imports-then-all assert_equals: reg.updateViaCache updated expected "all" but got "imports"
FAIL register-with-updateViaCache-imports-then-none assert_equals: reg.updateViaCache updated expected "none" but got "imports"
FAIL register-with-updateViaCache-all-then-undefined assert_true: New worker installing expected true got false
FAIL register-with-updateViaCache-all-then-imports assert_true: New worker installing expected true got false
FAIL register-with-updateViaCache-all-then-all assert_equals: reg.updateViaCache updated expected "all" but got "imports"
FAIL register-with-updateViaCache-all-then-none assert_equals: reg.updateViaCache updated expected "none" but got "imports"
FAIL register-with-updateViaCache-none-then-undefined assert_true: New worker installing expected true got false
FAIL register-with-updateViaCache-none-then-imports assert_true: New worker installing expected true got false
FAIL register-with-updateViaCache-none-then-all assert_equals: reg.updateViaCache updated expected "all" but got "imports"
FAIL register-with-updateViaCache-none-then-none assert_equals: reg.updateViaCache updated expected "none" but got "imports"
Harness: the test ran to completion.
